# api/openapi.json
{
    "components": {
        "schemas": {
...
            "api.DetectRpmsResponse": {
                "properties": {
                    "found": {
                        "description": "List of rpm names found in given repositories",
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "missing": {
                        "description": "List of rpm names not found in given repositories",
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
            "api.RepositoryRpmCollectionResponse": {
                "properties": {
                    "data": {
                        "description": "List of rpms",
                        "items": {
                            "$ref": "#/components/schemas/api.RepositoryRpm"
                        },
                        "type": "array"
                    },
                    "links": {
                        "$ref": "#/components/schemas/api.Links"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/api.ResponseMetadata"
                    }
                },
                "type": "object"
            },
...
            "api.SearchRpmResponse": {
                "properties": {
                    "package_name": {
                        "description": "Package name found",
                        "type": "string"
                    },
                    "package_sources": {
                        "description": "List of the module streams for the package",
                        "items": {
                            "$ref": "#/components/schemas/api.PackageSourcesResponse"
                        },
                        "type": "array"
                    },
                    "summary": {
                        "description": "Summary of the package found",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "errors.ErrorResponse": {
                "properties": {
                    "errors": {
                        "items": {
                            "$ref": "#/components/schemas/errors.HandlerError"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "RhIdentity": {
                "in": "header",
                "name": "x-rh-identity",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {},
        "description": "The API for the repositories of the content sources that you can use to create and manage repositories between third-party applications and the [Red Hat Hybrid Cloud Console](https://console.redhat.com). With these repositories, you can build and deploy images using Image Builder for Cloud, on-Premise, and Edge. You can handle tasks, search for required RPMs, fetch a GPGKey from the URL, and list the features within applications.\n",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "title": "ContentSourcesBackend",
        "version": "v1.0.0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/repositories/{uuid}/rpms": {
            "get": {
                "description": "List RPMs in a repository.",
                "operationId": "listRepositoriesRpms",
                "parameters": [
                    {
                        "description": "Repository ID.",
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Number of items to include in response. Use it to control the number of items, particularly when dealing with large datasets. Default value: `100`.",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Starting point for retrieving a subset of results. Determines how many items to skip from the beginning of the result set. Default value:`0`.",
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Term to filter and retrieve items that match the specified search criteria. Search term can include name.",
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Sort the response based on specific repository parameters. Sort criteria can include `name`, `url`, `status`, and `package_count`.",
                        "in": "query",
                        "name": "sort_by",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.RepositoryRpmCollectionResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List Repositories RPMs",
                "tags": [
                    "rpms"
                ]
            }
        },
...
        "/rpms/names": {
            "post": {
                "description": "This enables users to search for RPMs (Red Hat Package Manager) in a given list of repositories.",
                "operationId": "searchRpm",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.ContentUnitSearchRequest"
                            }
                        }
                    },
                    "description": "request body",
                    "required": true,
                    "x-originalParamName": "body"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/api.SearchRpmResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unsupported Media Type"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Search RPMs",
                "tags": [
                    "rpms"
                ]
            }
        },
        "/rpms/presence": {
            "post": {
                "deprecated": true,
                "description": "This enables users to detect presence of RPMs (Red Hat Package Manager) in a given list of repositories.",
                "operationId": "detectRpm",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.DetectRpmsRequest"
                            }
                        }
                    },
                    "description": "request body",
                    "required": true,
                    "x-originalParamName": "body"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.DetectRpmsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unsupported Media Type"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Detect RPMs presence",
                "tags": [
                    "rpms"
                ]
            }
        },
...
        "/snapshots/rpms/names": {
            "post": {
                "description": "This enables users to search for RPMs (Red Hat Package Manager) in a given list of snapshots.",
                "operationId": "searchSnapshotRpms",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.SnapshotSearchRpmRequest"
                            }
                        }
                    },
                    "description": "request body",
                    "required": true,
                    "x-originalParamName": "body"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/api.SearchRpmResponse"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "415": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unsupported Media Type"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/errors.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Search RPMs within snapshots",
                "tags": [
                    "rpms"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "https://console.redhat.com/api/content-sources/v1.0/"
        }
    ]
}